hieunv_delete

+++ src/main/java/com/capstone/booking/service/impl/PaymentMethodsServiceIml.java
@@ -9,6 +9,7 @@ import com.capstone.booking.repository.OrderRepository;
 import com.capstone.booking.repository.PaymentMethodsRepository;
 import com.capstone.booking.service.PaymentMethodsService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -46,8 +47,12 @@ public class PaymentMethodsServiceIml implements PaymentMethodsService {
 
 
     @Override
-    public void delete(long id) {
+    public ResponseEntity<?> delete(long id) {
+        if (!methodsRepository.findById(id).isPresent()) {
+            return new ResponseEntity("Id already exists", HttpStatus.BAD_REQUEST);
+        }
         methodsRepository.deleteById(id);
+        return new ResponseEntity("Delete Successful", HttpStatus.OK);
     }
 
     @Override
